+++ actions/destroy.py
@@ -1,7 +1,7 @@
 import os
 from lib import action
-from python_terraform import IsFlagged
-
+from dda_python_terraform import IsFlagged
+from dda_python_terraform.terraform import TerraformCommandError
 
 class Destroy(action.TerraformBaseAction):
     def run(self, plan_path, state_file_path, target_resources, terraform_exec,
@@ -23,6 +23,7 @@ class Destroy(action.TerraformBaseAction):
         """
         os.chdir(plan_path)
         self.terraform.terraform_bin_path = terraform_exec
+        self.set_semantic_version()
         self.terraform.targets = target_resources
         return_code, stdout, stderr = self.terraform.destroy(
             plan_path,
@@ -30,6 +31,8 @@ class Destroy(action.TerraformBaseAction):
             var=variable_dict,
             state=state_file_path,
             force=IsFlagged,
-            capture_output=False
+            capture_output=False,
+            raise_on_error=False
         )
+
         return self.check_result(return_code, stdout, stderr, return_output=True)