+++ actions/init.py
@@ -1,6 +1,7 @@
 import os
 from lib import action
-from python_terraform import IsFlagged, IsNotFlagged
+from dda_python_terraform import IsFlagged, IsNotFlagged
+from dda_python_terraform.terraform import TerraformCommandError
 
 
 class Init(action.TerraformBaseAction):
@@ -19,12 +20,14 @@ class Init(action.TerraformBaseAction):
         """
         os.chdir(plan_path)
         self.terraform.terraform_bin_path = terraform_exec
+        self.set_semantic_version()
 
         return_code, stdout, stderr = self.terraform.init(
             plan_path,
             backend_config=backend,
             capture_output=False,
-            upgrade=IsFlagged if upgrade else IsNotFlagged
+            upgrade=IsFlagged if upgrade else IsNotFlagged,
+            raise_on_error=False
         )
 
         return self.check_result(return_code, stdout, stderr)